Thin out to stub driver.
The code in dev-als has been partly converted from the the veml7700 driver. At least at first glance, the veml7700 driver is not well written and the conversion does not appear careful.

The efforts in dev-als showed that the PCB is ok but the calculated values from the driver are rubbish.

This stub is in to allow other dev to progress and retain the known good bit of the code. The guts can be re-visited later.